requirements.txt
pytest>=8.0
requests>=2.31
python-dotenv>=1.0
pytest-html>=4.0
pytest-ordering>=0.6
